pyproject.toml
[project]
name = "mardownstyler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "markdown-it-py>=4.0.0",
    "weasyprint>=66.0",
]
